# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "E-commerce application (Product: Shoes)",
        "title": "MyShoo API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/admin/dashboarddata/{range}": {
            "get": {
                "security": [
                    {
                        "BearerTokenAuth": []
                    }
                ],
                "description": "Get dashboard data (for a given date range)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Analytics"
                ],
                "summary": "Get dashboard data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Range",
                        "name": "range",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start Date",
                        "name": "sd",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End Date",
                        "name": "ed",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetDashBoardDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.SME"
                        }
                    }
                }
            }
        },
...
        "/admin/top-brands": {
            "get": {
                "security": [
                    {
                        "BearerTokenAuth": []
                    }
                ],
                "description": "Get top brands",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Analytics/Top_Performers"
                ],
                "summary": "Get top brands",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "l",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.TopBrandsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.SME"
                        }
                    }
                }
            }
        },
        "/admin/top-models": {
            "get": {
                "security": [
                    {
                        "BearerTokenAuth": []
                    }
                ],
                "description": "Get top models",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Analytics/Top_Performers"
                ],
                "summary": "Get top models",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "l",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.TopModelsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.SME"
                        }
                    }
                }
            }
        },
        "/admin/top-products": {
            "get": {
                "security": [
                    {
                        "BearerTokenAuth": []
                    }
                ],
                "description": "Get top products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Analytics/Top_Performers"
                ],
                "summary": "Get top products",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "l",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.TopProductsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.SME"
                        }
                    }
                }
            }
        },
        "/admin/top-sellers": {
            "get": {
                "security": [
                    {
                        "BearerTokenAuth": []
                    }
                ],
                "description": "Get top sellers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Analytics/Top_Performers"
                ],
                "summary": "Get top sellers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "l",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.TopSellersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.SME"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.DashboardData": {
            "type": "object",
            "properties": {
                "cancelled_order_count": {
                    "type": "integer"
                },
                "coupon_discounts": {
                    "type": "number"
                },
                "net_original_value": {
                    "type": "number"
                },
                "net_sale_value": {
                    "type": "number"
                },
                "order_count": {
                    "type": "integer"
                },
                "returned_order_count": {
                    "type": "integer"
                },
                "sale_value_after_cancellation_and_returns": {
                    "type": "number"
                },
                "users_registered": {
                    "type": "integer"
                }
            }
        },
...
        "entities.SalePerDay": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "sale": {
                    "type": "number"
                }
            }
        },
...
        "response.GetDashBoardDataResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "sale_per_day": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.SalePerDay"
                    }
                },
                "sales_report": {
                    "description": "need update",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entities.DashboardData"
                        }
                    ]
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "response.SME": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "response.TopBrands": {
            "type": "object",
            "properties": {
                "brand_id": {
                    "type": "integer"
                },
                "brand_name": {
                    "type": "string"
                },
                "net_mrp_value_sold": {
                    "type": "number"
                },
                "net_sale_amount": {
                    "type": "number"
                },
                "quantity_sold": {
                    "type": "integer"
                },
                "seller_name": {
                    "type": "string"
                }
            }
        },
        "response.TopBrandsResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "topBrands": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.TopBrands"
                    }
                }
            }
        },
        "response.TopModels": {
            "type": "object",
            "properties": {
                "brand_name": {
                    "type": "string"
                },
                "model_id": {
                    "type": "integer"
                },
                "model_name": {
                    "type": "string"
                },
                "net_mrp_value_sold": {
                    "type": "number"
                },
                "net_sale_amount": {
                    "type": "number"
                },
                "quantity_sold": {
                    "type": "integer"
                },
                "seller_name": {
                    "type": "string"
                }
            }
        },
        "response.TopModelsResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "topModels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.TopModels"
                    }
                }
            }
        },
        "response.TopProducts": {
            "type": "object",
            "properties": {
                "brand_name": {
                    "type": "string"
                },
                "colour_variant_id": {
                    "type": "integer"
                },
                "current_mrp": {
                    "type": "number"
                },
                "current_sale_price": {
                    "type": "number"
                },
                "model_name": {
                    "type": "string"
                },
                "net_mrp_value": {
                    "type": "number"
                },
                "net_sale_amount": {
                    "type": "number"
                },
                "quantity_sold": {
                    "type": "integer"
                },
                "seller_name": {
                    "type": "string"
                }
            }
        },
        "response.TopProductsResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "topProducts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.TopProducts"
                    }
                }
            }
        },
        "response.TopSellers": {
            "type": "object",
            "properties": {
                "net_mrp_value_sold": {
                    "type": "number"
                },
                "net_sale_amount": {
                    "type": "number"
                },
                "quantity_sold": {
                    "type": "integer"
                },
                "seller_id": {
                    "type": "integer"
                },
                "seller_name": {
                    "type": "string"
                }
            }
        },
        "response.TopSellersResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "topSellers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.TopSellers"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerTokenAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}